+++ docs/source/changelog.rst
@@ -8,6 +8,20 @@ The page attempt to keep a clear list of breaking/non-breaking changes and new f
    :local:
    :backlinks: none
 
+v0.13.0
+-----------
+Bug Fixes
+###########
+* Renamed `base.search` `category` parameter to `search_category` to avoid conflict with filters
+* Fixed the way IDs are retrieved
+* Fixed a bug where saving files and medias would not stream the response
+* Fixed issue where the front page would sometimes fail to parse
+
+New Features
+##############
+* `File` and `Addon` now have a `location` value that contains the location list of the entity
+* `File.save`, `Addon.save` and `Media.save` can now take `chunk_size` as a keyword parameter to define the size of the chunks to stream in
+
 v0.12.0
 -------
 Bug Fixes